requirements.txt
discord.py
torch
transformers
mineflayer
python-dotenv